added new commands MoveToAndExtend, as well as MoveToAndStow, added Features to FollowCurve and RobotPath
MoveToAndExtend waits till the robot is a set distance from the target position then begins extending, while extending it sets a speed cieling, it ends when it is at position, and finished extending to the set position. MoveToAndStow immediately starts stowing the elevator and sets a speed cieling while doing so, it returns to max speed once it has finished stowing. added additional features to FollowCurve to set a speed ceiling, and to directly get the remaining path length. RobotPath.getRemaingPathLength() now adds the robots distance from the closest curve position to the remaining length to provide more accurate results when the path is almost complete and allow it to reccorect more accurately at the end of a path.